adding simple fio profile for IO run and yaml file to accept input
Adding a simple fio profile which can run a quick IO without taking
much time as the default profile "fio-mixed.job" file takes.
And added a yaml file to accept the inputs like disk on which it
can run the IO and a job profile file.
altered the fiotest.py to run with yaml file.

Signed-off-by: Naresh Bannoth <nbannoth@in.ibm.com>