Testsuite to validate perf_event_subsystem
It covers an archive of perf_events.h files covers the history of interface ,fuzzer tool to hit the
extremes of perf_event interface,tests for the PAPI library and various perf_event related tools.
This series of test is meant to validate that the perf_event subsystem is working.

Signed-off-by: Shriya <shriyak@linux.vnet.ibm.com>