Perf test suite to validate the perf interface
This is a series of test is meant to validate that the perf_event subsystem is working.
It supports features like perf fuzzer, that tries to hit the extremes of the perf_event interface and find bugs in the kernel
perf_event_h, an archive of perf_event.h files over the history of the interface.Useful when figuring out when a feature was added.
It also supports various perf_event related tools and also tests PAPI library if it is supported.

Signed-off-by: Shriya <shriyak@linux.vnet.ibm.com>