Test to verify 24x7 hardware counters
24*7 performance counters is supported on Power VM platform.
Script will cover basic testing of events in 24x7 counters that belong to different
domains like core,chip,virtual CPU home core, etc.To monitor an event, the user must
specify the required parameters for the event,which include its name,
the chip/core on which to monitor the event as well as the domain.
Added the check for VCPU, added failure info for all failures, optimized the code.

Signed-off-by: Shriya <shriyak@linux.vnet.ibm.com>