Energy mangement: Tuned tests
“Tuned” in RHEL is tuning daemon for automatically tuning the system via the use of tuning profile. It can also be configured
to react to changes to improve performance of the server and also to make system settings persistent.
tuned-adm is a command line tool that provides a number of different profiles to improve performance in a number of specific use cases
tuned-adm includes the ability to run any command as part of enabling or disabling a tuning profile.
Several predefined profiles are already included. You can even create your own profile, either based on one of the existing ones by copying it or make a completely new one. The distribution provided profiles are stored in subdirectories below /usr/lib/tuned
and the user defined profiles in subdirectories below /etc/tuned. If there are same profiles on both places, user defined profiles have precedence.

Signed-off-by: Shriya <shriyak@linux.vnet.ibm.com>