Cpuhotplug
Hotplug the cpus in random order by selecting random core.
The script will select random core from the list, picks the random cpus and does
offline and online in random order for that core.Make sure that all cores
are online.Script will set smt to maximum value,later it
works in random order to offline and online the cpu for 100 iterations.

Signed-off-by: Shriya <shriyak@linux.vnet.ibm.com>