Using full namespace for tests
Some tests support input as both full namespace or specifying
namespace id with -n option. But that is applicable only when
there is one namespace. When more than one namespace is active,
input using -n is not supported. Full namespace name needs to be
given. Changed that in such tests.

Signed-off-by: Narasimhan V <sim@linux.vnet.ibm.com>