Throughput test for SMT folding in presence of swizzle
SMT folding is achieved through hotplug.To verify this test, run ebizzy workload when SMT is off
and when SMT is on. Observe the records generated for single thread for 60s for one cpu.
Expected is single thread performance is better then multi thread hence it is proved by comparing
the records generated.

Signed-off-by: shriyak <shriyak@linux.vnet.ibm.com>