To validate quad-level frequency transitions
On Power 9, when one of the cpu's frequency is set to maxmimum or higher than other cpus beloging
to same quad,then complete quad will be set to max frequency irrespective of the frequency on
other cpus.This test will set the governor to userspace, and randomly set the frequency on each cpu.
Maximum frequency set for each quad is recorded  and it will be verfied by running perftool on each cpu
to monitor the cycles, which will give frequency count as same as maximum frequency recorded for that
quad.

Signed-off-by: shriyak <shriyak@linux.vnet.ibm.com>