Adding a sleep time for EEH recovery
EEH Recovery is not similar for all adapters. For some adapters,
specifically multipath, we see that the adapter needs some more
time to recover after the message "Notify device driver to resume"
on the dmesg.

There is no reliable way to determine this extra time required,
nor a way to determine the recovery. So, a sleep time of 10s is
introduced.

Signed-off-by: Narasimhan V <sim@linux.vnet.ibm.com>