Throughput test for SMT folding in presence of swizzle
SMT folding is achieved through hotplug.To verify this test, the script runs ebizzy workload
when SMT is off and when SMT is on. Observe the records generated for single thread for 60s on
one cpu.Expected is single thread performance is better then multi thread and it is proved by
comparing the records generated.To observe better results, the script disables all the idle states
before running ebizzy using cpupower tool.

Signed-off-by: Praveen K Pandey <praveen@linux.vnet.ibm.com>
	     : Shriya Kulkarni	<shriyak shriyak@linux.vnet.ibm.com>