Convert dmesg switch '-c' to '-C'
dmesg buffer is cleared '-c', as well as with '-C'. Whereas the latter
does not prints the contain of dmesg before clearing it, while '-c'
prints the dmesg buffer before clearing it. We are not interested in
the dmesg before the test run's so convert the switch from '-c' to '-C'.

Signed-off-by: Kamalesh Babulal <kamalesh@linux.vnet.ibm.com>