kernel/kernbench: Use olddefconfig to build kernel
Upstream Linux Kernel commit fb16d8912db5 ("kconfig: replace
'oldnoconfig' with 'olddefconfig', and keep the old name as an alias")
replaces 'oldnoconfig' with 'olddeconfig' and for compatibility reasons
maintains an alias, which is targeted to be removed by Kernel v4.19.

This patch switches over to 'olddefconfig', this also helps in avoiding
the depreciation usage warning about 'oldnoconfig'.

Signed-off-by: Kamalesh Babulal <kamalesh@linux.vnet.ibm.com>