ras/kprobes: optprobes test
By default the optprobes is enabled for the kernel functions and this
test disables the optprobes using the '/proc/' knob available to
disable optrobes, while planting a kprobe. By disabling optprobes, the
kprobe mechanism fallbacks to un-optimizes probes. It reuses the kprobes
kernel module already build to test un-optimized version of kprobe.

Signed-off-by: Kamalesh Babulal <kamalesh@linux.vnet.ibm.com>