Fix the raw code for POWER9 generic events
The generic events on POWER8 and POWER9 differ in raw code and hence the
event name remains, only raw code changes. This fix will update the raw
code for POWER9 events like
LLC-stores,cache-references,L1-dcache-loads,L1-dcache-load-misses etc.
Example for POWER8 'LLC-stores' the raw code is : 0x17080 and for POWER9
it is :0x16880

The raw code value for each event is obtained under sysfs directory :
/sys/bus/event_source/devices/cpu/events
for POWER8 and POWER9.

Signed-off-by: Shriya <shriyak@linux.vnet.ibm.com>
Signed-off-by: Kamalesh Babulal <kamalesh@linux.vnet.ibm.com>