Changing bonding interface parameter in yaml
Bonding test needed input interfaces as host_interfaces, which is
now changed into bond_interfaces.
This is needed, since when we run multiple tests with bonding,
and if they need to run on the bond interface created, then the
parameter name on bonding test and other tests need to be
different.

Signed-off-by: Narasimhan V <sim@linux.vnet.ibm.com>