Split the htx_network test into start, check, stop
Split the htx_network test into start, check, stop, so that the
test can be start to make HTX run in background and other tests
can be run in parallel, and HTX can be stopped later.
test_check can be used whenever necessary, to check if HTX is
still running in the background.

Signed-off-by: Narasimhan V <sim@linux.vnet.ibm.com>