Warning when not sure of the failure in nvmetest
There is no clear way to find out the firmware version from the
firmware file that we are downloading. So, the script was using
the firmware file name as firmware version also, which is a
workaround. When we fail the test based on that, it is a
false failure. So, we are exiting with a warning in this case.

Signed-off-by: Narasimhan V <sim@linux.vnet.ibm.com>