Correct two problems in fiotest.py
Runs with Filesystems require a different fio command than runs without
filesystems. This patch creates a seperate command for the no_fs case.
It also corrects the path for the --filename variable in the fs case.

This change may expose a known issue with fio:
(see: https://github.com/axboe/fio/issues/740)
There is temporary code to make this an avocado warning instead of an
error because the test still completes.

We should declare variables before test cancels and goes to clean up
because clean up expects fs_create and lv_create. Current code throws an
error when btrfs is not supported.

Signed-off-by: Kyle Mahlkuch <kmahlkuc@linux.vnet.ibm.com>