Use `--prefix=/usr` by default for glibc
The `glibc.py` test specifies a _prefix_ of `self.build_dir` when doing
`./configure`.

This is mostly innocuous unless a `make install` is performance, which
is not done by `glibc.py`.

It is an issue, however, when doing a `make check` because some glibc
tests are unable to find the system's `libgcc_s.so.1` and FAIL.
There is an obscure reference to this known issue at
https://sourceware.org/glibc/wiki/Testing/Testsuite#Known_testsuite_failures

Thus, default _prefix_ to `/usr`.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>