ndctl: Remove "all" from disable/enable test case
Removing "all" is valid because if create-namespace fails, the pmem lib will
throw an error. And if one of the namespace disable fails, it will still throw
an error. So we got all the cases covered. This "all" is redundant.

If in future, ndctl treats "all" to behave differently for zero namespaces, we
will again have to change the test cases. Future proof it.

Signed-off-by: Santosh S <santosh@fossix.org>