ibmvnic: increase wait time when pinging after client failover
It takes time for ARP-request packets to resolve when trying to send packets to
the peer after a client failover. This fix allows the vNIC to wait 50
seconds for a replay for each packet sent.

Additionally, using write_file_or_fail instead of write_file to catch errors
when writing to a file.

Signed-off-by: Cristobal Forno <cris.forno@ibm.com>
Co-developed-by: Dany Madden <danymadden@us.ibm.com>