virt-net: test interface every time DLPAR remove and add occurs
Currently, the interface is only tested after X amount of DLPAR remove/add
operations. This fix adds an ip addr, brings the link up, and pings the peer
with the interface after every dlpar add. On remove, it checks to see if the
interface is still present.

Additionally, the interface is removed via chhwres on the HMC instead
of using the drmgr on the LPAR.

Signed-off-by: Cristobal Forno <cforno12@linux.ibm.com>