Add support to use upstream fsprogs while testing
Modify fs/xfstests.py to pull and make fsprogs like e2fsprogs,
xfsprogs and btrfsprogs from upstream. This feature is useful
when testing upstream kernels.

To test for upstream, we can use the run_type option in the
yaml. We also need to specify the upstream URLs of the fsprogs
we want to install, eg:

...
upstream:
  run_type: 'upstream'		# Test with upstream fsprogs
  xfsprogs_url: 'https://git.kernel.org/pub/scm/fs/xfs/xfsprogs-dev.git'
  e2fsprogs_url: 'https://git.kernel.org/pub/scm/fs/ext2/e2fsprogs.git'
  btrfsprogs_url: 'https://git.kernel.org/pub/scm/linux/kernel/git/kdave/btrfs-progs.git'
distro:
  run_type: 'distro'		# Test with distro provided fsprogs

Signed-off-by: Ojaswin Mujoo <ojaswin@linux.ibm.com>