Use distro as a default target for toolchain component tests
A number of tests under toolchain directory has support for
both distro and upstream as a target (as a YAML input).

Lot of these toolchain tests have a dependency on compilers and
libraries installed on the system where they are executed.

In order to avoid test errors due to mismatch in levels being
tested vs what is available on the system, switch to using
distro as the default target for all the applicable tests.

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>