When running some tests on unsupported OS (Ubuntu), the confusing failure messages arise caused by the fact that some variables are not initialized in setUp but used in the tearDown.
In memory/pmem* and ndctl, let's check if we have corresponding attributes set
before trying to use them.

In selinux-tests, we can initiate the problematic variable to None in the beginning
and then use it only if it is not empty.

ras/lparstat is just missing explicit test cancelation for the OS other than
rhel/suse.

Signed-off-by: Denis Silakov <dsilakov@gmail.com>