memory/ndctl: Correct iniparser header dependency
Commit 26b9f92(enable support for meson build) unconditionally fails
the tests when using upstream as a target. Remove this restriction
for releases that ship iniparser package.

Additionally the code in autotools_build_system unconditionally
copies iniparser headers to required directory. This copy operation
fails on releases that ship these headers. Add a check for such releases
and skip the copy operation to avoid test failures.

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>