Done rework for va_test.c
With older implementations, MAP_CHUNK_SIZE is fixed to 16 GB, as a
result fix value, On low memory the test case is getting failed with
ENOMEM error message where it clearly indicates that not enough memory
is present in the system, which results in a few va_test cases failing.
The above changes will dynamically adjust the MAP_CHUNK_SIZE value
according to the amount of space in the system.

Signed-off-by: Samir Mulani <samir@linux.vnet.ibm.com>