This patch makes sure the slave interfaces are cleaned up after deleting creating bond on peer lpar
To achieve this below commands are used in cleanup functions
ip addr flush dev interface-name
ip addr add ip-address/netmask dev interface-name
ip link set dev interface-name up

For flushing the ip address and making the interfaces up at peer lpar, public ip is used.
when bond interface is deleted and the slave ip's are flushed out , ssh connection will be lost and script gets stuck
To avoid this using public ip in cleanup to flush the addresses and make the interfaces up.

Removing "ifup and ifdown" commands, as ifup and if down are not supported in latest rhel9 versions. And using bringup() and bringdown() util finctions in cleanup.

Signed-off-by: Tasmiya Nalatwad <tasmiya@linux.vnet.ibm.com>