[patch 3/4] ras/ras.py: Don't use absolute path for rtas_errd
The code uses absolute path of rtas_errd in run_cmd. While this
is okay, the code will not work if a user has compiled code from
source and installed the binaries in different directory.

Remove the hard coded value so that the binary file can be picked
up based on PATH directive.

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>