vnic: Get backing devices for specific slot
Previously, the function get_backing_devices returned a string
where every line showed the backing devices of a particular vnic
device. This was a problem because, if there were more than one vnic
device on the lpar, then the calling functions would get a list of
irrelevant backing devices.

Therefore, fix get_backing_devices so that it returns the backing
devices of a specific vnic device(slot).

Signed-off-by: Nick Child <nick.child@ibm.com>