ras/ras_ppcutils: Move lparstat tests into ppcutils script
lparstat.py test has support for distro as a target. Adding
upstream as a target to this test will require duplication of
code which is already part of ras_ppcutils.py. Since lparstat
tool is part of powerpc-utils package, consolidating lparstat
tests as a part of ras_ppcutils will avoid the duplication of code.

This patch removes lparstat.py test and moves the corresponding
tests to ras_ppcutils.py. With this movement upstream as a
target support will be automatically available for lparstat.

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>