ras/ras_ppcutils: Move numascore tests into ppcutils script
lparnumascore.py test has support for distro as a target. Adding
upstream as a target to this test will require duplication of
code which is already part of ras_ppcutils.py. Since lparnumascore
tool is part of powerpc-utils package, consolidating lparnumascore
tests as a part of ras_ppcutils will avoid the duplication of code.

This patch removes lparnumascore.py test and moves the corresponding
tests to ras_ppcutils.py. With this movement upstream as a
target support will be automatically available for lparnumascore.

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>