Skip Rcutorture test when module cannot be loaded
In distros such as RHEL check_kernel_config may not be able to detect
if CONFIG_RCU_TORTURE_TEST is enabled or not. Try loading and unloading
the module to check if test is runable else skip it. This is to accommodate
RHEL not placing kernel config file in the system while building upstream
kernel.

Signed-off-by: Ayush Jain <ayush.jain3@amd.com>