Resolve duplicate debug msg in memory/fork_mem test
Here below are the logs of duplicate msg "mmaping xx anonynous bytes"

[stdlog] 2023-01-19 06:20:26,564 avocado.utils.process DEBUG| [stdout] mmaping 10485760 anonymous bytes
[stdlog] 2023-01-19 06:20:26,564 avocado.utils.process DEBUG| [stdout] PID 31747 touching 160 pages
[stdlog] 2023-01-19 06:20:26,564 avocado.utils.process DEBUG| [stdout] mmaping 10485760 anonymous bytes
[stdlog] 2023-01-19 06:20:26,564 avocado.utils.process DEBUG| [stdout] PID 31753 touching 160 pages
[stdlog] 2023-01-19 06:20:26,564 avocado.utils.process DEBUG| [stdout] mmaping 10485760 anonymous bytes

Becuase, stdout is streamed to some file, the return character ('\n') is
not able to flush the stdout stream and on each time fork is called, the
parent log are printed again.

Also this patch resolve some compiler time warnings.

Signed-off-by: Tarun Sahu <tsahu@linux.ibm.com>
Reported-by: Geetika <geetika@linux.ibm.com>