xfstests: Add args parameter feature support
Add a simplified way of doing xfstests by passing an "args" parameter
such that if each fs_type is provided with "args" argument, then it is
directly passed to "./check" script.
This also move check for fs_to_test at the beginning.

This simplifies a lot of things for parameter handling. In later patches
we will remove options like "group", "test_range".

Signed-off-by: Ritesh Harjani (IBM) <ritesh.list@gmail.com>