Move htx_test to workload folde
The Hardware Test Executive (HTX) is a suite of test tools for
hardware validation of OpenPOWER system. The HTX stress tests the
system by exercising all or selected hardware components concurrently
to uncover any hardware design flaws and
hardware-hardware or hardware-software interaction issues.

As this test is usually run as a wrokload with other tests moving this
to the new wrokload tests folder.

Signed-off-by: Geetika <geetika@linux.ibm.com>