perf/perf_hv_gpci: Handle non root user addition
This script allows a non root user to execute perf commands and
has checks in place to add such a non root user if it does not
exist.

If addition of non root user fails, the script logs a warning
message. Extend this logic to run_cmd() function as well to
avoid runtime error if user does not exist.

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>