perf/perf_pmu: Handle non root user creation
This script allows a non root user to execute perf commands and
has checks in place to add such a non root user if it does not exist.

There are places where certain action is taken assuming such a non
root user exist. Update this logic to avoid runtime error if user
does not exist.

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>