Removed netiface module usage from irqbalance.py
This test case is common for both net and block device.
netiface module is not required for block device test runs
and hence it is failing if we run for block device if we
don't install it before running the test.
So I removed it and for net I have used /sys/class/net
directory output for getting all the interfaces details.
so effect on any devices i,e net/block with this changes.
it will work as it was earlier.

Signed-off-by: Naresh Bannoth <nbannoth@in.ibm.com>