Rename test 'test' to 'test_basic' in order to run singly
On running 'avocado run numactl.py:test' all the tests are being executed.
As per docuemntation, when the test references are about Instrumented Tests,
Avocado will find any Instrumented test that starts with the reference,
like a “wildcard”.

Thus renaming 'test' to 'test_basic' in order to be able to run it singly.

Signed-off-by: Geetika <geetika@linux.ibm.com>