perf: add test to run json events
This patch adds a new test case to run all the perf pmu events.
Also compares the event list and event code between perf
command and json files from kernel source.
Also added upstream as a target so the test can be
run on a host booted with upstream kernel.

Signed-off-by: Disha Goel <disgoel@linux.ibm.com>