EEH test for vnic interfaces
Adding test to trigger EEH for vnic interfaces from vios. The test does the following:

change the active logport to the original logport
login to the vios and find the respective backing device
trigger EEH from vios
check for the failover on the OS side to conclude if the eeh was successful or not.

Signed-off-by: Vaishnavi Bhat <vaishnavi@linux.vnet.ibm.com>