Adapting the new changes made in htx setup/configuration.
The build_net multisystem command introduced is for automatic configuration
of network interfaces on both host and peer.
The command set's the interfaces with some local net id's and pings the interfaces.
The command also starts the htx deamon creates/selects and activates net.mdt
for the network devices.
No need to explicitely start the htx deamon or create/select or activate net.mdt file.

Also removing the code that was supporting the old htx network topology and
made compatible to new configuration changes.

Signed-off-by: Tasmiya Nalatwad <tasmiya@linux.vnet.ibm.com>