perf/perf_metric.py: add test for CPI STALL json feature
Patch adds a new function to test CPI_STALL_RATIO metricgroup by checking
if its present in perf list or not. If metricgroup is present, collect
counter statistics using perf stat and check if all stall metrics are
present in the output. Also check for multiplexing of events should not
happen.

Test for metricgroup events fails when performance information collection
property is not enabled in hmc since internally it uses hv_24x7 events to
collect statistics. Added a check to cancel the test in such cases.

Signed-off-by: Disha Goel <disgoel@linux.ibm.com>