ftrace_log_fix
Ftrace tests were ending with a pass even if there were failures
in the tests. This was due to the code reading the older
snapshot of the log file for matching the 'not ok' string.
Therefore, modify logic to paste the log to a different
raw file and read the file to match for 'not ok' string.

Signed-off-by: Akanksha J N <akanksha@linux.ibm.com>