.github/workflows/xfstests: Fix ubuntu and opensuse version requirements
For ubuntu:latest
Latest pip3 mandates to use virtual managed environment for pip
packages. However since this is mainly automation and we know what we ar
doing so pass the flag ("--break-system-packages") to force pip to install
the package system wide.

For opensuse/leap:latest
Latest avocado-framework=100 and PyYAML=6.0.2 are incompatible.
So move avocado-framework to 100.1 and PyYAML=6.0.1 which are still
compatible.

Signed-off-by: Ritesh Harjani (IBM) <ritesh.list@gmail.com>