Making disk_info.py script platform independent
currently the script will run only on power server and gets cancel
on other platform. Due to recent issue in python AI(platform) it is
not returning the 'ppc64' when trying to retrive the processor type.
And script can also be ran any linux machine irrespective of any platform
so removed this dependency as this is a generic script.

Signed-off-by: Naresh Bannoth <nbannoth@in.ibm.com>