kselftest: result variable to be accessed universally
result variable, which stores the command output of selftest,
was being run either from test() or run_cmd(). This is based
on the component for which it was being run. This patch makes
it to be accessed from inside test() irrespective of where it is
run, to access the output for post process.

Additionally, packages to be installed are also fixed to be in
line with latest verions of Ubuntu.

Signed-off-by: Narasimhan V <16101053+narasimhan-v@users.noreply.github.com>