Fixed uses of showing warning in avocado misc tests.
In older versions of Avocado, the self.warn method was used to display warnings,
and we adapted this approach when the API supported it. However, in the latest
version of Avocado, warnings are now displayed using self.log.warn(), as warning
messages have been incorporated into the logging class.
We have adopted the same API in multiple test cases where we previously used the older method for displaying warnings.

Signed-off-by: Samir Mulani <samir@linux.vnet.ibm.com>

Fixed PEP8 compliance FAIL issue.

Ran the autopep8 -i .py file to fix the style issue.

Signed-off-by: Samir Mulani <samir@linux.vnet.ibm.com>